Change span for new names to `def_site()` by default
This will be changed in the near future: def_site() is only used
when this crate is compiled in 'nightly' mode, where `call_site()`
will be used in stable.

Note that in almost all cases, the hygiene won't make a difference
in practice. The names are already choosen to not conflict with
anything.