Guard unstable `proc_macro` features behind "nightly` feature
This is most of the work for #13. This crate now compiles in two modes:

With "nightly" feature enable, `proc_macro_diagnostic` and
`proc_macro_span` are enable to improve our error messages.

Without the "nightly" feature, we don't use those unstable features
and instead build a workaround for the lack of unstable features
(something like a poly fill). Most of the workaround can be found in
the `diag` module.

The crate doesn't compile with the stable compiler quite yet. There
are still three feature gates activated. But those three features
will be stabilized for Rust 2018 and they are easy to remove by only
rewriting a few lines in the whole crate.