Add two examples that fail to compile (to showcase error messages)
The examples are a bit awkward for some reason: examples are compiled when
one executes `cargo test`. This is nice, but we actually want those examples
to fail. We can't just exclude them, so we only compile the interesting
part when the `fail` cfg is set.

I hope to find a better solution for this...