Fix late-bound lifetime warning
PR #35 introduced a warning in one of the examples. We were passing all
generic parameters to the inner function, including lifetimes. But
passing late-bound lifetimes won't be allowed in the future. This
commit simply ignores lifetime parameters. AFAICT this should never be
a problem. Passing lifetimes explicitly hardly ever makes sense. The
only case I can think of where the changes of this commit could break
anything is in this function:

    fn foo<'a>() -> &'a i32 { ... }

But functions with an early-bound lifetime (return type) and no
late-bound ones are extremely rare and special. So for now we can
ignore them. In particular, determining which lifetime parameters are
early- and which are late-bound is extremely difficult.

Lastly, const parameters are not yet correctly passed to the inner
function. This needs to be fixed in the future, when const parameters
are a thing.