Fix #23: methods with `self` receiver for `FnMut` proxy type
Previously the generated code would be incorrect as `self` (by value)
must be declared `mut` to call it as `FnMut`. To avoid special casing,
the call for all `Fn*` proxies is `({self})(args)` now (`{x}` is
basically the identify function). That way it's not necessary to bind
`self` as mut.

For more information why this works, you can read this:
https://bluss.github.io/rust/fun/2015/10/11/stuff-the-identity-function-does/