Use lifetime elision in impl headers for generated impls
Fixes #22

Instead of generating `impl<'a, T> Foo for &'a T` we now generate
`impl<T> Foo for &T`. This works in newer versions of Rust. See
https://github.com/rust-lang/rust/issues/15872 for more information.

This change makes a lot of other code useless. In particular, we don't
need to find a special lifetime parameter name anymore. Everything was
adjusted to remove this useless code.