Add `?Sized` relaxation to impl header whenever possible
It's only impossible when there is `Self` by value in in parameter or
return position in any method. This commit does not yet check for
`Self` in non-receiver parameters as this crate cannot deal with that
yet anyway.

Furthermore, a `where Self: Sized` bound on methods does not help in
that we still cannot add the `?Sized` relaxation. This is related to
issue #11.