Split UI tests into two test functions and change logic in CI script
The problem is that we only want to run compile-fail tests for a single
compiler version (the current stable) on CI. This is because compiler
errors change over time and we can only support one version at a time.

Before this commit, this was solved by simply disabling all tests for
non-stable compilers. But this had a few downsides:

- You had no way of running these tests with a nightly compiler locally,
  which is a valid thing to do.
- We also disabled the compile-pass tests which don't have the problem
  and should always work on all compilers.
- I want to add `1.37` (MSRV) to our CI matrix. There, the "stable"
  logic would break, as that's stable, too.

So now the only check is in the CI script itself.